\name{log2.ratios}
\alias{log2.ratios}
\title{Extracting log2 ratios}
\description{
  This function extracts the log2 ratios from either an
  \code{\link[limma:malist]{MAList}} object or a \code{\link[snapCGH:SegList]{SegList}} object.
}

\usage{log2.ratios(x)}

\arguments{
  \item{x}{An object of class \code{\link[limma]{MAList}} or \code{\link[snapCGH:SegList]{SegList}}}
}

\author{Mike Smith}

\keyword{datasets}
