\name{log2ratios}
\alias{log2ratios}
\title{Extracting log2 ratios}
\description{
  This function extracts the log2 ratios from either an
  \code{\link[limma:malist]{MAList}} object or a \code{\link[snapCGH:SegList]{SegList}} object.
}

\usage{log2ratios(x)}

\arguments{
  \item{x}{An object of class \code{\link[limma:malist]{MAList}} or
    \code{\link[snapCGH:SegList]{SegList}}}
}

\author{Mike Smith}

\keyword{datasets}
