\name{mod.chrominfo}
\alias{mod.chrominfo}
\docType{data}
\title{Chromosomal information}
\description{
  This dataset provides information about the length and centromere of the 24 chromosomes which form the human genome. The length and centromere are given in base pairs.
}
\usage{data(mod.chrominfo)}
\format{
  A data frame with 24 observations on the following 3 variables.
  \describe{
    \item{\code{chrom}}{the chromosome}
    \item{\code{length}}{the length of the chromosome}
    \item{\code{centromere}}{the position of the centromere}
  }
}

}
\source{
  This information was taken from the aCGH package of Jane Fridlyand and Peter Dimitrov
}

\keyword{datasets}
