\name{non.zero.length.distr.non.tiled}
\alias{non.zero.length.distr.non.tiled}
\docType{data}
\title{Empirical distribution of segment lengths in non-tiled regions with copy number gains or losses}
\description{
  This file contains the empirical distribution of segment lengths (of untiled regions and whose state mean indicates that they correspond to regions of copy number gain or loss) derived by applying the DNAcopy segmentation scheme (Olshen et al., 2004) to an unpublished breast cancer dataset. Instead of using the physical length of the segments we calculate the lengths as a proportion of the length of the untiled region.
}
\usage{data(non.zero.length.distr.non.tiled)}
}
\source{
 The empirical distribution was derived using an unpublished breast cancer dataset.
}
\keyword{datasets}
