\name{zero.length.distr.tiled}
\alias{zero.length.distr.tiled}
\docType{data}
\title{Empirical distribution of segment lengths in tiled regions with no copy number gains or losses}
\description{
  This file contains the empirical distribution of segment lengths (of tiled regions and whose state mean indicates that they correspond to regions of no copy number gain or loss) derived by applying the DNAcopy segmentation scheme (Olshen et al., 2004) to an unpublished breast cancer dataset. Instead of using the physical length of the segments we calculate the lengths as a proportion of the length of the tiled region.
}
\usage{data(zero.length.distr.tiled)}
\source{
  The empirical distribution was derived using an unpublished breast cancer dataset.
}
\keyword{datasets}
